Change request context from Span to a wrapper object RequestContext (#7)
In order to substitute TChannel's native trace reporting, we want to override its tracing hook, as well as its RequestContextProvider. That requires the request context from this module to be extensible. Currently it only stores the `Span`, so this change introduces a `RequestContext` class that wraps the span and can be further extended.