Add response_handler_hook
As a software engineer, I'd like to be able to
add custom tags to the span depending of the response.
This PR adds an ability to set response_handler_hook using
set_response_handler_hook method of RequestsPatcher.
response_handler_hook should be a method that accepts response
and span as positional arguments which will be provided by send_wrapper
just after the response is received.

These changes also:

 - Add base Patcher

 - Refactor opentracing_instrumentation/client_hooks/requests.py

 - Add tests for the requests

 - Simplify the singleton decorator

 - Fix strict_redis.reset_patches method